pyproject.toml
[tool.poetry]
name = "dynamic distributed computing"
version = "0.1.0"
description = ""
authors = ["Joy2225 <archishman9892@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dill = "^0.3.8"
pycryptodome = "^3.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
